🐛 Enable -Wall -Wextra and fix all library warnings
WARNFLAGS was empty, so the library had never been compiled with
warnings on. Turns them on and fixes every warning that originates in
src/EZ-Template/ and include/EZ-Template/:

- Sign-compare (int vs size_t) on index loops against .size(): loop
  variables changed to std::size_t where the index is never used
  signed elsewhere in the function (drive.cpp x9, exit_conditions.cpp,
  pid_tasks.cpp, purepursuit_math.cpp x6, set_odom_pid.cpp x3,
  util.cpp x2). Where the variable is also compared against 0/negative
  elsewhere in the same function (pid_tuner.cpp's column, util.cpp's
  screen_print/get_rest_of_the_word loops) or feeds a printf %i, the
  .size() side is cast to int instead, matching the cast-to-int
  convention already used elsewhere in this codebase for pp_index.
- Unused variables removed: fake_x (purepursuit_math.cpp, half of a
  rotation pair where only the y component is used), short_error_sgn
  (purepursuit_math.cpp, computed but never read), flip_a
  (set_odom_pid.cpp, theta flipping is done by flip_angle_target
  instead), slew_on in the QAngle overload of pid_swing_set
  (set_swing_pid.cpp, the value it fed into is discarded since the
  overload calls into the 5-arg version, which recomputes the same
  value itself), analog_left_value/analog_right_value
  (user_input.cpp, read and then never referenced; the joysticks are
  re-read a few lines later instead).
- delete-non-virtual-dtor on pros::v5::Imu* in Drive::~Drive(): every
  pointer in all_imus is allocated as exactly `new pros::Imu(...)`
  (never a derived type), so the delete is safe; pros::Imu is a
  vendored header so the destructor itself can't be fixed. Suppressed
  locally with a scoped #pragma GCC diagnostic and a comment, rather
  than a project-wide -Wno- flag.

No warnings appeared from vendored headers (include/pros, include/okapi,
include/liblvgl), so the -isystem reclassification described as a
fallback wasn't needed.

Adds a warnings-clean CI job that rebuilds just the library archive
(`make library`, which the Makefile already exposes and which skips
src/main.cpp and src/autons.cpp) with -Werror layered on top. Those two
excluded files are compiled by the default build but aren't part of the
published library, and src/main.cpp currently has one sign-compare
warning of its own (ez_motor_temperatures) that's out of scope here, so
a -Werror job over the whole project would fail on a file this PR
doesn't touch. -Werror is intentionally kept out of the Makefile itself
so a contributor's local build never hard-fails on a compiler version
difference. The job reuses the exact toolchain from the existing build
job's LemLib/pros-build@v5.1.0 (arm-gnu-toolchain 14.3.1, verified
against the pinned image's Dockerfile) by invoking that same pinned
Docker image directly with an overridden entrypoint, since the action
itself takes no input for extra build flags or a custom target.

+++ src/EZ-Template/drive/pid_tuner.cpp
@@ -118,7 +118,7 @@ void Drive::pid_tuner_print() {
   if (!pid_tuner_on) return;
 
   // Ensure that the user column is within the size of the pid tuner
-  if (column > used_pid_tuner_pids->size() - 1)
+  if (column > (int)used_pid_tuner_pids->size() - 1)
     column = 0;
 
   double kp = used_pid_tuner_pids->at(column).consts->kp;
@@ -209,7 +209,7 @@ void Drive::pid_tuner_iterate() {
   // Up / Down for Rows
   if (master.get_digital_new_press(pid_tuner_pageRight)) {
     column++;
-    if (column > used_pid_tuner_pids->size() - 1)
+    if (column > (int)used_pid_tuner_pids->size() - 1)
       column = 0;
     pid_tuner_print();
   } else if (master.get_digital_new_press(pid_tuner_pageLeft)) {
